EdgeContainer/modules/opcmodul/requirements.txt
azure-iot-device~=2.7.0

